requirements.txt
icalendar==5.0.4
openrgb_python==0.3.0
requests==2.31.0
